new | modalControl.js
모달 창 제어 함수 분리 및 스크립트 파일 정리
